08/11 - service layer updates @jondoetsch27

+++ src/main/kotlin/com/jdd/springboot/kotlin/service/impl/PlayerServiceImpl.kt
@@ -1,5 +1,7 @@
 package com.jdd.springboot.kotlin.service.impl
 
+import com.jdd.springboot.kotlin.exception.DuplicatePlayerException
+import com.jdd.springboot.kotlin.exception.PlayerNotFoundException
 import com.jdd.springboot.kotlin.model.Player
 import com.jdd.springboot.kotlin.repository.PlayerRepository
 import com.jdd.springboot.kotlin.service.PlayerService
@@ -17,20 +19,36 @@ class PlayerServiceImpl : PlayerService {
     }
 
     override fun readPlayer(playerName: String): Player {
-        return playerRepository.findById(playerName).get()
+        return if (playerRepository.findById(playerName).isPresent) {
+            playerRepository.findById(playerName).get()
+        } else {
+            throw PlayerNotFoundException()
+        }
     }
 
     override fun createPlayer(newPlayer: Player): Player {
-        return playerRepository.save(newPlayer)
+        return if (playerRepository.existsById(newPlayer.playerId)) {
+            throw DuplicatePlayerException()
+        } else {
+            playerRepository.save(newPlayer)
+        }
     }
 
     override fun updatePlayer(newPlayer: Player): Player {
-        playerRepository.deleteById(newPlayer.playerId)
+        if (playerRepository.existsById(newPlayer.playerId)) {
+            playerRepository.deleteById(newPlayer.playerId)
+        } else {
+            throw PlayerNotFoundException()
+        }
         return playerRepository.save(newPlayer)
     }
 
     override fun deletePlayer(oldPlayer: Player): Player {
-        playerRepository.deleteById(oldPlayer.playerId)
+        if (playerRepository.existsById(oldPlayer.playerId)) {
+            playerRepository.deleteById(oldPlayer.playerId)
+        } else {
+            throw PlayerNotFoundException()
+        }
         return oldPlayer
     }
 }
